requirements.txt
# DesignAsCode - Python Dependencies
# Core dependencies for inference

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Model Hub and Transformers
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Training
trl>=0.7.0
peft>=0.4.0

# Data Processing
datasets>=2.11.0
pydantic>=2.0.0

# Semantic Search and Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
# Note: Use faiss-gpu instead for GPU acceleration

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# API and HTTP
openai>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.60.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.10.0

# Logging and Monitoring
tensorboard>=2.12.0
wandb>=0.13.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter support (optional)
jupyter>=1.0.0
ipython>=8.0.0

# HTML rendering
playwright>=1.40.0
html5lib>=1.1
lxml>=4.9.0
